Final Update
This is the finalized looks of this project. quite responsive (i hope). and also probably from today's onward i will be making fewer projects, since i want to start with Javascript and must first learn the basic fundamental. but will probably make a JS projects when i can, and when i reached that point.